+++ cnccode/data_classes/DBEItemType.inc.php
@@ -22,7 +22,7 @@ class DBEItemType extends DBCNCEntity
     const showInCustomerReview   = "showInCustomerReview";
     const sortOrder              = "sortOrder";
     const allowGlobalPriceUpdate = "allowGlobalPriceUpdate";
-
+    const showStockLevels        = "showStockLevels";
 
     /**
      * calls constructor()
@@ -81,6 +81,13 @@ class DBEItemType extends DBCNCEntity
         );
         $this->addColumn(
             self::allowGlobalPriceUpdate,
+            DA_BOOLEAN,
+            DA_NOT_NULL,
+            null,
+            false
+        );
+        $this->addColumn(
+            self::showStockLevels,
             DA_BOOLEAN,
             DA_NOT_NULL,
             null,